refactor/#3: Split the function file into separate functions
Closes #3
