+++ .gitignore
@@ -52,3 +52,23 @@ Mkfile.old
 dkms.conf
 .idea
 cmake-build-debug/
+
+### CMake ###
+CMakeLists.txt.user
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Testing
+Makefile
+cmake_install.cmake
+install_manifest.txt
+compile_commands.json
+CTestTestfile.cmake
+_deps
+CMakeUserPresets.json
+
+noto
+
+### CMake Patch ###
+# External projects
+*-prefix/
