--- Makefile
@@ -1,4 +1,0 @@
-build:
-	@echo "[+] Building..."
-	@gcc -Wall main.c functions.c -o noto
-	@echo "[+] Done!"
