+++ src/headers/create_entry.h
@@ -1,0 +1,1 @@
+int create_entry(char *str);
