+++ src/headers/create_new_db.h
@@ -1,0 +1,1 @@
+void create_new_db();
