+++ src/headers/delete_line.h
@@ -1,0 +1,3 @@
+#include <stdio.h>
+
+void delete_line(FILE *srcFile, FILE *tempFile, int line);
