+++ src/headers/generate_hash.h
@@ -1,0 +1,1 @@
+char *generate_hash();
