+++ src/headers/get_path_to_db.h
@@ -1,0 +1,1 @@
+char *get_path_to_db();
