+++ src/headers/main.h
@@ -1,0 +1,8 @@
+#include "create_entry.h"
+#include "create_new_db.h"
+#include "delete_line.h"
+#include "generate_hash.h"
+#include "get_path_to_db.h"
+#include "remove_entry.h"
+#include "show_all.h"
+#include "show_entry.h"
