+++ src/headers/show_all.h
@@ -1,0 +1,1 @@
+void show_all();
