+++ src/headers/show_entry.h
@@ -1,0 +1,1 @@
+int show_entry(char *str);
